requirements.txt
tensorflow-gpu==1.14.0
numpy==1.16.4
keras==2.3.0
pandas==1.1.5
scikit-learn==0.22.2
dgl==0.9.0
dgllife==0.3.0
torch==1.9.0